- zzz.R now performs a check to ensure filehash is available.
- checkDictionaryStatus now examines the value of the global option
  tikzMetricsDictionary and if a path is specified it will create/load
the dictionary specified in there instead of making a disposable one in
tempdir(). This allows the user to specify a permanent dictionary in
their .Rprofile that will remember metrics in between R sessions. With a
permanent dictionary test #12 now executes in ~0.7 seconds. All other
tests execute in lest than a tenth of a second.

- The use of metrics dictionary files does not have many safety switches
  at this stage. Case in point: if a permananent dictionary is specified
in .Rprofile it MUST be specified as an absolute path. filehash chokes
when dealing with relative paths for some reason. All the code dealing
with dictionaries could probably benefit from some safeties.

But protecting the ignorant is not very high on the todo list :p