Added a test using XeLaTeX
Because the preview package does not work with XeLaTeX and I could
not get the f-ing options to work correctly in the test suite and
ghostscript will not combine the pdf because of the fonts, the XeLaTeX
test is in a separate file which gets run before all the regular tests.

Some error handling has been added for when metrics calculations fail,
the latex log is printed out and processing is stopped.

Header, footer and documentclass options have been added using the global
options() function.  In the future this should be implemented in a
separate environment .tikzOptions with a set function tikzOptions() and an
access function getTikzOptions().

A function setTikzDefaults() has been added to tikz.R and is exported.