Added stub code for raster functions
R 2.11.0 introduced support for plotting raster images and capturing the
state of a graphics device as a raster image.  Stub code has been added
for both of these device functions.  Without it, R 2.11.0 would crash if
a user attempts to use a function such as rasterImage()- error messages
are now returned instead.

Code related to this new functionality is protected inside C
preprocessor directives which check for R_GE_version to be 6 or greater.
This prevents older versions of R, which do not include the new
functions, from getting confused.