Added better version number handling
`git describe` is a nice command that formats version numbers using a
combination of the most recent git tag, the number of commits since that
tag and a SHA1 partial of the most recent commit.  However, the Version:
field of the DESCRIPTION file won't allow such strings as version
numbers.

A new file, inst/GIT_VERSION has been added to hold git version strings
and will be referenced by getTikzDeviceVersion() which will distribute
the number throughout the code.

New script, updateVersion.sh, updates inst/VERSION with the output of
`git describe`, or its first argument if one was passed.

The README.md and DESCRIPTION files will always contain the version
number of the next release.