Added `/usr/texbin` to list of places to search
On UNIX-like operating systems, people are likely to have a varient of TeX Live
installed.  The R.app GUI on OS X has a nasty problem of ignoring environment
variables set by things such as the MacTeX installer which place `/usr/texbin`
on the Terminal search path.

This patch adds `/usr/texbin/pdflatex` to the end of the list of places
tikzDevice will search for a compiler if the operating system is UNIX.