Started working on raster output.
Raster output is pretty tricky business-- unless there is some clever way to
include raster data in a plaintext `.tex` file that TikZ knows about.

Current idea is to call back to R and output a raster using the PNG graphics
device.  Then the tikzDevice will add an `\includegraphics` call to the TeX
output.

This commit contains work towards a proof-of-concept:

- A new R function, `tikz_writeRater`, has been added to deviceUtils.R.  This
  function is unexported (not visible to users) and will handle writing the PNG.

- A new counter, rasterFileCount, has been added to `tikzDevDesc` in
  `tikzDevice.h`.  This counter will be incremented each time a raster is
  created.  Raster files will use this counter to generate file names such as
  'foo_raster_1.png` and `foo_raster_2.png` if the main file is `foo.tex`.

- The C level callback to `tikz_writeRaster` has been implemented in
  `tikzDevice.c`.  The callback currently passes required variables to R.
  However, it does not yet do anything with the return value.