Worked on R-level implementation of raster output.
The R callback function for creating rasters has been brought to a working
prototype stage.  One significant problem is that png() does not behave
consistantly across platforms.  On OS X quartz() must be used so anti-aliasing
can be disabled (perhaps this should be a user-controlled option?). On Windows
png() appears to drop a row or column of data or otherwise suffers from some
sort of off-by-one error.

The C implementation has been re-worked to return a list of red, green, blue,
alpha values instead of a vector of C unsigned ints that have been coerced to an
INTSXP as this made no sense and was unusable at the R level.