Re-add XeLaTeX font variant test
Re-added the XeLaTeX test for font variant support. The graphics tests may now
use an optional `graph_options` parameter which may be set to a list. If
present, this list is passed to `options()` at the beginning of the test.
Changes are reverted at the end of the test.

The XeLaTeX font variant test may fail on platforms other than OS X if the
Zapfino font is not available.