Add a fallback to getTikzDeviceVersion
For some reason, newer versions of Windows are not copying GIT_VERSION from the inst directory which is causing getTikzDevice version to throw a warning when this file cannot be found. This patch causes tikzDevice to use the version number in the DESCRIPTION file as a fallback. The number is prepended with a tilde so we know what happened if we see it in a bug report.