Remove code for ?text example in vignette
The code in this example contained Unicode characters which cause `R CMD check`
to throw warnings under R 2.13.0. Adding `\usepackage[utf8]{inputenc}` does not
help as `pdflatex` then throws warnings concering Unicode characters for which
no glyphs are defined.

Given the poor support for Unicode in `pdftex` it would take a lot of work to
sort this out before the release of R 2.13.0. Therefore, it is being removed
for now although the output is retained for comparison.

Closes GH-22.