Fix declaration of line end variable
A variable containing line ending info was mistakingly declared as containing
line join info. Probably wasn't causing any problems as both types are
enumerations of the same three integer values, but it never hurts to be
consistent.

This bugfix brought to you by the asskicking powers of the Clang compiler.