Duplicate TikZ sources
Straight up clone of the TikZ source to form the basis of an experimental
METAPOST device. All `TikZ` function prefixes have been substituted for `MetaP`
and some routines have been flagged as static so that the package will compile.

The resulting `metapost` device is able to start up and generate a `plot(1,1)`
that compiles to PDF. Of course, this is just `tikz` producing TikZ code under
a different name, but it still works!