Reimplemented clipping for MetaPost
MetaPost works slightly different than TikZ. A "new frame" is popped onto a
stack using `pushcurrentpicture`. When drawing has concluded in a frame, it can
be clipped and then drawing returns to the old frame with `popcurrentpicture`.

The push and pop operations are analogous to TikZ's begin and end scope and the
difference is that the clip happens at the end of the scope rather than at the
beginning.