Added support for MetaPost text
MetaPost has an odd way of quoting TeX strings:

   btex Tex stuff here etex

Or:

   "Tex stuff here"

Both produce slightly different results. btex/etex looks a bit better. They
both have the potential to produce some wierd escaping problems---with TikZ, a
node was ended by `}` which was a TeX special character so it had to be escaped
inside strings anyway. Using a double quote or the word `etex` presents a more
difficult escaping problem.

Once again, style info is not added---aside for selection of bold/italic
typefaces.