Class for MetaPost colors
Well, technically ConTeXt colors, but they will be used by MetaPost. This patch
adds a `metap_color` object that contains a name, a RGB value and possibly a
"spot color" name that can be used to isolate the color on a seperate channel.

Also included is a `format` method for turning `metap_color` objects into
ConTeXt code.