Added support for maintaining lists of colors
A little convoluted, but this code adds support for maintaining hashed lists of
RGB colors. There is one list attached to each MetaPost device and a global
list that they all share.

The `define_metap_color` function can be used to add a `metap_color` object to
any of these lists. Color lists can only contain one entry for a given RGB
value. When a value is redefined, the new `metap_color` object replaces the old
one in the list. However, some attributes of the old object, such as it's name,
are copied into the new object so any references to the old names in the
MetaPost code can be found and re-written later.