Link MetaPost color code with C level device
The device code now calls up to the R level for a color name that goes with a
given RGB value. If none is found in the device color list or the global color
list, a new entry is added to the device color list and the name is returned.