Restore color to MetaPost drawing routines
Unlike TikZ, it appears that MetaPost cannot fill and draw with different
colors in the same operation. So, all drawing routines now define a define a
path which is stored in the variable `p`. A `fill` or `draw` operation is then
executed on `p` to apply the correct colors. The draw operations could be
extended to handle line type, width and dashing using MetaPost pen definitions.