Fixed some errors in device initialization
The draw code was being triggered on a fill which was causing a stray
`begingroup` to get printed along with draw commands. Surprisingly, MetaPost
had no problem with this. Also, need to explicitly set the bounds of the image
in case the DrawStyle decides not to put in a fill because it is transparent.