Compare graphs using ImageMagick
Output PDFs are compared against a "standard" graph using the `compare` utility
from ImageMagick. The two input PDFs are converted to PNG at a density of 300
dpi and then diffed. The AE (absolute error) metric is used which returns the
number of pixels that were different.

Need to do some cross platform testing to establish an acceptable threshold of
difference.