Add a tikzNode annotation
This function will create a node with the following pieces:

  \node[opts] (name) at (x,y) {contents};

If any one of those pices is NULL, it will be omitted. If `contents` is NULL, a
'coordinate' will be generated instead of a node.

Due to this last behavior, `tikzCoord` is now implemented as a friendly wrapper
for `tikzNode`.