Ensure tikzNode only uses scalar values.
It would be great to vectorize tikzNode so that a bunch of nodes could be
created with one function call. However, handling the `opts`, `name` and
`contents` arguments in that case will require some careful consideration so we
enforce scalar `x` and `y` values and leave vectorization for the future.