Remove newlines from text nodes
It appears that the sequence '%\n' may cause some problems in Windows where
'\n' is not translated into a Windows-specific endline '\r\n' (I assumed it
was). The sequence '%\n' is commonly used in LaTeX macros to prevent newlines
creating unwanted spaces. There is really no point to having '%\n' in the code,
so I am pulling it out.

For more info:

  https://groups.google.com/d/topic/tikzdevice/x8RsfdhAepk/discussion