Restructure blocks into a hierarchical list
Using the Pandoc function `hierarchicalize`, Block lists can be re-structured
such that they are lists of lists grouped by heading. This makes it easy to
fulfill the Rd requirement that all text be an argument to a `section` or
`subsection` command.

`md2news.hs` now produces a usable Rd file for the first time.