Twaddle some magic numbers in tikzDevice.c
Set the values of `yLineBias`, `xCharOffset` and `yCharOffset` to be the same
as values used in R's PDF device. This should help even out the spacing of
margin text in base graphics. Prior to this change, text on the Y axis was
recieving about half the leading used for text on the X axis.

Interestingly, this has no effect on non-margin text in base graphics. Text in
Grid graphics appears to be entirely unaffected.

Fixes issue #49.