Define an enumeration to describe drawing operations
Currently, the decision to draw a path, fill it or both is made by calculating
an integer "code". This system is overly complicated and does not even work.

A better system is to define an enumeration with values that are sucessive
powers of two. This forms a set of "flags" that can be combined using a bitwise
or. The presence of a given flag can be tested using a bitwise and.