Retire SetColor from background fill
The background fill is now generated using `TikZ_DefineColor` instead of
`SetColor`.

The background is now created with the `path` command rather than the `fill`
command and has the `use as bounding box` key set. Setting this key should free
TikZ from the burden of computing the bounding box for the graphic when TeX is
compiling.