Defer starting new tikzpicture environments
The creation of a new `tikzpicture` environment, or "new page" in R graphics
terms, is now deferred until output has been generated to occupy that page.
This change builds on the machinery created for issue #45 that avoids the
production of empty clipping scopes.

The `firstPage` member of the `tikzInfo` structure is no longer necessary and
has been removed.

Fixes #12.