Revise vignette cover page
The vignette cover page was laid out using a lot of weird calculations and
positioning hacks that probably make no sense to anyone but me.

This revision cleans up the code and moves a few pieces of info such as the
authors, date and package version onto the cover page.

Commands such as `\author` and `\title` that get inserted into the PDF metadata
have been cleaned up.