getDocumentPointsize: fix regexp
The previous regexp would match any number that was followed by either
"p" or "t" and use this as the document point size.  This led to

    \documentclass[a4paper]{artice}

being interpreted as a 4pt document.  As a result, plotted figures
wound up with grossly scaled font nodes (scale=3 or similar).

We instead want to only match digits at the beginning of a word,
followed by the string "pt".